+++ components/navbar/index.tsx
@@ -47,7 +47,7 @@ const Navbar: React.FC = () => {
         <Link key={name} href={route} passHref>
           <Button
             as="a"
-            colorScheme={isRoute(route) ? 'red' : undefined}
+            colorScheme={isRoute(route) ? 'yellow' : undefined}
             fontWeight={isRoute(route) ? 'bold' : 'normal'}
             variant="ghost"
           >