+++ theme/index.ts
@@ -1,10 +1,6 @@
-import {
-  extendTheme,
-  theme as defaultTheme,
-  ThemeOverride,
-} from '@chakra-ui/react';
+import { extendTheme, theme as defaultTheme } from '@chakra-ui/react';
 
-export default extendTheme(<ThemeOverride>{
+export default extendTheme({
   components: {
     Heading: {
       baseStyle: {
@@ -15,7 +11,22 @@ export default extendTheme(<ThemeOverride>{
     Link: {
       variants: {
         link: {
-          color: 'red.400',
+          color: 'yellow.200',
+        },
+      },
+    },
+    Modal: {
+      baseStyle: {
+        dialog: {
+          bgColor: 'gray.900',
+        },
+        overlay: {
+          '@supports (backdrop-filter: blur(4px))': {
+            backdropFilter: 'blur(4px)',
+          },
+          '@supports (-webkit-backdrop-filter: blur(4px))': {
+            WebkitBackdropFilter: 'blur(4px)',
+          },
         },
       },
     },
@@ -32,6 +43,9 @@ export default extendTheme(<ThemeOverride>{
 
   styles: {
     global: {
+      ':root': {
+        colorScheme: 'dark',
+      },
       '::selection, ::-moz-selection': {
         bgColor: 'whiteAlpha.400',
       },
@@ -63,7 +77,7 @@ export default extendTheme(<ThemeOverride>{
         pointerEvents: 'none',
       },
       '#nprogress .bar': {
-        bgGradient: 'linear(to-r, whiteAlpha.400, red.400)',
+        bgGradient: 'linear(to-r, whiteAlpha.400, yellow.200)',
         h: '2px',
         left: 0,
         pos: 'fixed',