+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.1.0](https://github.com/Fadhilamadan/fadhilamadan.com/compare/v2.0.2...v2.1.0) (2021-10-02)
+
+
+### Features
+
+* **api:** add throw error for sentry ([0e4b42f](https://github.com/Fadhilamadan/fadhilamadan.com/commit/0e4b42f46728178c8e208e2bf25d6acbcfcf0af4))
+* **sentry:** add default config ([4f16b31](https://github.com/Fadhilamadan/fadhilamadan.com/commit/4f16b319eb3a1290725dbdca30bb24d45136f43b))
+
 ### [2.0.2](https://github.com/Fadhilamadan/fadhilamadan.com/compare/v2.0.1...v2.0.2) (2021-09-29)